+++ lite/src/core/treemap.hpp
@@ -1,5 +1,5 @@
 
-const uint32 the_treemap_path_size = 32;
+const int64 the_treemap_path_size = 32;
 
 template< typename Key, typename Value, class TreeIndex = Index<Key> >
 class TreeMap
@@ -216,7 +216,7 @@ public:
     {
         the_page_size = 0;
     }
-    TreeMap(uint32 page_size)
+    TreeMap(int64 page_size)
     {
         init(page_size);
     }
@@ -249,7 +249,7 @@ public:
         Node* pos = the_root;
         while(pos)
         {
-            int32 dir = the_index.compare((Key&)key, pos->key);
+            int64 dir = the_index.compare((Key&)key, pos->key);
             if(dir)
                 pos = pos->leaves[0 < dir];
             else
@@ -257,19 +257,19 @@ public:
         }
         return Find();
     }
-    void page_size(uint32 page_size)
+    void page_size(int64 page_size)
     {
         assert(!the_page_size);
         init(page_size);
     }
-    uint32 page_size()
+    int64 page_size()
     {
         return the_page_size;
     }
-    uint64 size()
+    int64 size()
     {
         assert(the_page_size);
-        uint64 size = the_tail - the_tail_page->data;
+        int64 size = the_tail - the_tail_page->data;
         Page* page = the_head_page;
         while(page->next)
         {
@@ -294,7 +294,7 @@ public:
         Node* pos = the_root;
         while(pos)
         {
-            int32 dir = the_index.compare((Key&)key, pos->key);
+            int64 dir = the_index.compare((Key&)key, pos->key);
             if(dir)
                 pos = pos->leaves[0 < dir];
             else
@@ -308,7 +308,7 @@ public:
         Node* pos = the_root;
         while(pos)
         {
-            int32 dir = the_index.compare((Key&)key, pos->key);
+            int64 dir = the_index.compare((Key&)key, pos->key);
             if(dir)
                 pos = pos->leaves[0 < dir];
             else
@@ -333,7 +333,7 @@ public:
         Node* pos = the_root;
         while(pos)
         {
-            int32 dir = the_index.compare((Key&)key, pos->key);
+            int64 dir = the_index.compare((Key&)key, pos->key);
             if(dir)
             {
                 i++;
@@ -371,7 +371,7 @@ public:
         Node* pos = the_root;
         while(pos)
         {
-            int32 dir = the_index.compare((Key&)key, pos->key);
+            int64 dir = the_index.compare((Key&)key, pos->key);
             if(dir)
             {
                 i++;
@@ -410,7 +410,7 @@ public:
         Node* pos = the_root;
         while(pos)
         {
-            int32 dir = the_index.compare((Key&)key, pos->key);
+            int64 dir = the_index.compare((Key&)key, pos->key);
             if(dir)
             {
                 i++;
@@ -448,7 +448,7 @@ public:
         Node* pos = the_root;
         while(pos)
         {
-            int32 dir = the_index.compare((Key&)key, pos->key);
+            int64 dir = the_index.compare((Key&)key, pos->key);
             if(dir)
             {
                 i++;
@@ -487,11 +487,11 @@ public:
         assert(-1 != self_test(the_root));
         return true;
     }
-    uint32 erase(const Key& key, bool unique = true)
+    int64 erase(const Key& key, bool unique = true)
     {
         assert(the_page_size);
-        uint32 cnt;
-        uint32 erased = 0;
+        int64 cnt;
+        int64 erased = 0;
         do
         {
             cnt = erased;
@@ -504,7 +504,7 @@ public:
             Node* pos = the_root;
             while(pos)
             {
-                int32 dir = the_index.compare((Key&)key, pos->key);
+                int64 dir = the_index.compare((Key&)key, pos->key);
                 if(dir)
                 {
                     i++;
@@ -572,7 +572,7 @@ public:
                         Node* tail_pos = the_root;
                         while(tail_pos != tail)
                         {
-                            int32 dir = the_index.compare(tail->key, tail_pos->key);
+                            int64 dir = the_index.compare(tail->key, tail_pos->key);
                             if(dir)
                             {
                                 parent = &tail_pos->leaves[0 < dir];
@@ -766,7 +766,7 @@ private:
             }
         }
     }
-    void init(uint32 page_size)
+    void init(int64 page_size)
     {
         assert(page_size);
         the_page_size = page_size + 1;
@@ -843,6 +843,6 @@ private:
     Page* the_tail_page;
     Node* the_tail;
     Node* the_root;
-    uint32 the_page_size;
+    int64 the_page_size;
     TreeIndex the_index;
 };
